refactor: 카메라 기능을 app/Camera 페이지로 분리
